+++ src/entrypoints/content/modules/skill-bonus/index.ts
@@ -1,10 +1,20 @@
 import '@/entrypoints/content/modules/skill-bonus/index.css'
 
 import type { Module } from '@/entrypoints/content/common/types/module'
+import { querySelector, querySelectorAll } from '@/entrypoints/content/common/utils/dom'
 import { pages } from '@/entrypoints/content/common/utils/pages'
-import runPlayerDetail from '@/entrypoints/content/modules/skill-bonus/handlers/player-detail'
-import runPlayerList from '@/entrypoints/content/modules/skill-bonus/handlers/player-list'
 import metadata from '@/entrypoints/content/modules/skill-bonus/metadata'
+import { applySkillBonuses } from '@/entrypoints/content/modules/skill-bonus/utils'
+
+const runPlayerDetail = (): void => {
+  const element = querySelector('#mainBody .playerInfo')
+  if (element) applySkillBonuses([element])
+}
+
+const runPlayerList = (): void => {
+  const elements = Array.from(querySelectorAll('#mainBody > .playerList > .teamphoto-player'))
+  applySkillBonuses(elements)
+}
 
 const skillBonus: Module = {
   metadata,
